First cut of the code.
Junction is a base class for all junctions.

JunctionAll, JunctionAny and JunctionOne define user-facing classes and helpers (none(), one(), any() and all(), plus versions that force the library either to copy the data or to leave it uncopied) for creating junctions.

JunctionSortedStore stores a sorted copy of the elements of a junction where necessary.  JunctionPiggyBackStore accesses the elements when no copy is needed.

JunctionReverseComparisons provides (n == any({1, 2, 3})), given (any({1, 2, 3}) == n).

samples.cpp shows some simple ways of using junctions.  Start reading there, because this library is easier to use than it is to read.

testbed.cpp is what its name implies.